Elminated compilation warning about signed vs. unsigned comparison.
git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/trunk@4068 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ libs/csterr/struct/ddgcache.h
@@ -353,7 +353,7 @@ public:
 		if (n->next())
 			return n->next();
 		// Find next bucket with something in it.
-		while (--b > _tail && !_bucket[b] ){}
+		while (--b > (short)_tail && !_bucket[b] ){}
 		return b < 0 ? 0 : _bucket[b];
 	}
 };
